% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utility_functions.R
\name{aggregate_edges}
\alias{aggregate_edges}
\title{Graph To Undirected}
\usage{
aggregate_edges(graph)
}
\arguments{
\item{graph}{Graph to make undirected}
}
\value{
Undirected graph
}
\description{
Graph To Undirected
}
