% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{default_clustering}
\alias{default_clustering}
\title{Default Clustering}
\usage{
default_clustering(graph)
}
\arguments{
\item{graph}{The input graph}
}
\value{
cluster_fast_greedy clustering
}
\description{
Default Clustering
}
