% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{efficiency_signature}
\alias{efficiency_signature}
\title{Extract efficiency signature}
\usage{
efficiency_signature(graph, team_graph, weights = NULL, range_param = 3)
}
\arguments{
\item{graph}{Graph where team is part of}

\item{team_graph}{Subgraph with only team members}

\item{weights}{A string that specifies attribute representing weights of the graph}
}
\value{
A number representing an efficiency
}
\description{
Extract efficiency signature
}
