% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utility_functions.R
\name{friends_friends}
\alias{friends_friends}
\title{Friends friends subgraph}
\usage{
friends_friends(graph, vertices, n_steps)
}
\arguments{
\item{graph}{Igraph object}

\item{vertices}{List of ...}

\item{n_steps}{An int representing number of steps away from 'vertices'}
}
\value{
Igraph object with 'vertices' and 'n_step' neighburs
}
\description{
Create a subgraph from a graph with vertices surrounding
input vertices with neighbouring vertices n_steps away.
}
