% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{ideation_signature}
\alias{ideation_signature}
\title{Extract ideation signature}
\usage{
ideation_signature(graph, weights = "weights")
}
\arguments{
\item{graph}{Graph to analyze}

\item{weights}{A string, specifies column representing weights of the graph}
}
\value{
Igraph column with values of ideation
}
\description{
Extract ideation signature
}
