% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{silo_quotient}
\alias{silo_quotient}
\title{Extract Silo Quotient}
\usage{
silo_quotient(graph, team_graph)
}
\arguments{
\item{graph}{The input graph}

\item{membership}{Character vector, for each vertex it gives its community.}
}
\value{
A tibble with the teams silo quotients
}
\description{
Extract Silo Quotient
}
