% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{silo_signature}
\alias{silo_signature}
\title{Extract silo signature}
\usage{
silo_signature(
  graph,
 
    membership = igraph::cluster_fast_greedy(igraph::as.undirected(subgraphene))$membership,
  weights = NULL
)
}
\arguments{
\item{graph}{The input graph.}

\item{membership}{Numeric vector, for each vertex it gives its community. The communities are numbered from one.}

\item{weights}{If not NULL then a numeric vector giving edge weights.}
}
\value{
Silo signature measure (aka modularity)
}
\description{
This function is igraph::modularity renamed + default membership set to cluster_fast_greedy
}
