% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utility_functions.R
\name{to_undirected_edges}
\alias{to_undirected_edges}
\title{Edge data to undirected}
\usage{
to_undirected_edges(data, from, to)
}
\arguments{
\item{data}{A data.frame containing two columns representing edges of a graph}

\item{from}{A string that specifies column associated with \code{from} ids}

\item{to}{A string that specifies column associated with \code{to} ids}
}
\value{
data.frame with single row per edge of undirected graph with extra
  column, \code{edge_count}, representing number of occurances
}
\description{
Convert edge data to undirected edge data and count
number of occurances per undirected edge.
}
