% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{vertices_colors}
\alias{vertices_colors}
\title{Title}
\usage{
vertices_colors(graph, attribute, palette = "Set3")
}
\arguments{
\item{graph}{Graph to plot}

\item{attribute}{A string representing vertex attribute to set color by}

\item{palette}{A string representing ColorBrewer palette}
}
\value{
List of colors for the vertices
}
\description{
Title
}
