% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{vertices_sizes}
\alias{vertices_sizes}
\title{Title}
\usage{
vertices_sizes(graph, attribute, default_size = 12, scale_factor = 1.5)
}
\arguments{
\item{graph}{Graph to plot}

\item{attribute}{A string representing vertex attribute to set color by}

\item{default_size}{A number representing size of 0 or NA values}

\item{scale_factor}{A number to scale attribute value by}
}
\value{
List of sizes for the vertices
}
\description{
Title
}
