% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{vulnerability_signature}
\alias{vulnerability_signature}
\title{Extract Vulnerability Signature}
\usage{
vulnerability_signature(graph, membership, weights = NULL)
}
\arguments{
\item{graph}{The input graph.}

\item{membership}{Vertex attribute to cluster by.}

\item{weights}{Weight vector. If the graph has a weight edge attribute, then this is used by default.}
}
\value{
A tibble of vertices with high vulnerability score
}
\description{
Extract Vulnerability Signature
}
