% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{vulnerabilty_signature}
\alias{vulnerabilty_signature}
\title{Title}
\usage{
vulnerabilty_signature(
  graph,
  membership = igraph::cluster_fast_greedy(igraph::as.undirected(graph))$membership,
  weights = NULL
)
}
\arguments{
\item{graph}{The input graph.}

\item{membership}{Numeric vector, for each vertex it gives its community. The communities are numbered from one.}

\item{weights}{If not NULL then a numeric vector giving edge weights.}
}
\value{
A tibble of vertices with high vulnerability score
}
\description{
Title
}
