% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{ideation_signature}
\alias{ideation_signature}
\title{Extract Ideation Signature}
\usage{
ideation_signature(graph, weights = NULL)
}
\arguments{
\item{graph}{Graph to analyze}

\item{weights}{Weight vector. If the graph has a weight edge attribute, then this is used by default.}
}
\value{
Igraph column with values of ideation on a scale from 1 to 10
}
\description{
Extract Ideation Signature
}
