% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{ideation_signature}
\alias{ideation_signature}
\title{Extract ideation signature}
\usage{
ideation_signature(graph, weights = NULL)
}
\arguments{
\item{graph}{Graph to analyze}

\item{weights}{A string that specifies attribute representing weights of the graph}
}
\value{
Igraph column with values of ideation on a scale from 1 to 10
}
\description{
Extract ideation signature
}
