% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{influence_signature}
\alias{influence_signature}
\title{Extract influence signature}
\usage{
influence_signature(graph, weights = "weights")
}
\arguments{
\item{graph}{Graph to analyze}

\item{weights}{A string that specifies attribute representing weights of the graph}
}
\value{
Igraph column with values of influence
}
\description{
Extract influence signature
}
