% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{influence_signature}
\alias{influence_signature}
\title{Extract Influence Signature}
\usage{
influence_signature(graph, weights = NULL)
}
\arguments{
\item{graph}{Graph to analyze}

\item{weights}{A string that specifies attribute representing weights of the graph}
}
\value{
Igraph column with values of influence on a scale from 1 to 10
}
\description{
This is a fishy implementation of influence
}
