% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{plot_graph}
\alias{plot_graph}
\title{Plot an igraph object}
\usage{
plot_graph(
  graph,
  edge_width_by = "weights",
  vertex_color_by = NULL,
  vertex_size_by = NULL
)
}
\arguments{
\item{graph}{Graph to plot}

\item{edge_width_by}{A string, sets width of edge based on value from column in E(graph)}

\item{vertex_color_by}{A string, sets color vertex  based on value from column in V(graph)}

\item{vertex_size_by}{A string, size of vertex based on value from column in V(graph)}
}
\value{
Plotted graph
}
\description{
Create a plot of an igraph object and give vertex and edge
style based on graph attributes.
}
