% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{silo_signature}
\alias{silo_signature}
\title{Extract Silo Signature}
\usage{
silo_signature(graph, membership = default_clustering(graph), weights = NULL)
}
\arguments{
\item{graph}{The input graph.}

\item{membership}{Character vector, for each vertex it gives its community.}

\item{weights}{If not NULL then a numeric vector giving edge weights.}
}
\value{
Silo signature measure of graph (aka modularity)
}
\description{
This function is igraph::modularity renamed + default membership set to cluster_fast_greedy
}
