% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{vulnerabilty_signature}
\alias{vulnerabilty_signature}
\title{Extract Vulnerability Signature}
\usage{
vulnerabilty_signature(
  graph,
  membership = default_clustering(graph),
  weights = NULL
)
}
\arguments{
\item{graph}{The input graph.}

\item{membership}{Character vector, for each vertex it gives its community.}

\item{weights}{If not NULL then a numeric vector giving edge weights.}
}
\value{
A tibble of vertices with high vulnerability score
}
\description{
Extract Vulnerability Signature
}
