% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{vulnerabilty_signature}
\alias{vulnerabilty_signature}
\title{Extract Vulnerability Signature}
\usage{
vulnerabilty_signature(graph, membership, weights = NULL)
}
\arguments{
\item{graph}{The input graph.}

\item{membership}{Vertex attribute to cluster by.}

\item{weights}{If not NULL then a numeric vector giving edge weights.}
}
\value{
A tibble of vertices with high vulnerability score
}
\description{
Extract Vulnerability Signature
}
